Address review comments and clean up test resources
In this commit addresses the review comments are updated and I am also
adding a method to delete all the test resources because I am seeing
that some unrelated test that use the same resource types that I create
here are failing, likely due to some new resources being added in this
test.

Move the `assertResponseBundle()` method to FHIRServerTestBase. Remove
@author tag, and multiple other requested changes.

Also, move the integration test class to its own package.

Signed-off-by: luisgarcc <luisg.dev1@gmail.com>